Examples from Previous assignment and beginnings of view
Made an examples folder from previous assignment database stuff (all the
garbage that was in that last assignment DB folder) and made bare bones
html and php for view. Currently the sidebar directs user to the html
files, not php controller files. Changes I want to make should happen
tonight (?) but work on controller can be done now as I will only be
changing how it looks, not how it behaves to  the back end